moved sngp representer into probly.method.sngp
