+++ src/probly/method/sngp/__init__.py
@@ -2,9 +2,16 @@
 
 from __future__ import annotations
 
-from probly.lazy_types import TORCH_MODULE
+from probly.lazy_types import TORCH_MODULE, TORCH_SAMPLE
 
-from ._common import SNGPPredictor, register, sngp, sngp_traverser
+from ._common import (
+    SNGPPredictor,
+    SNGPRepresenter,
+    compute_categorical_sample_from_logits,
+    register,
+    sngp,
+    sngp_traverser,
+)
 
 
 ## Torch
@@ -13,8 +20,15 @@ def _(_: type) -> None:
     from . import torch as torch  # noqa: PLC0415
 
 
+@compute_categorical_sample_from_logits.delayed_register(TORCH_SAMPLE)
+def _(_: type) -> None:
+    from . import torch as torch  # noqa: PLC0415
+
+
 __all__ = [
     "SNGPPredictor",
+    "SNGPRepresenter",
+    "compute_categorical_sample_from_logits",
     "register",
     "sngp",
     "sngp_traverser",
