+++ src/probly/method/sngp/torch.py
@@ -2,11 +2,14 @@
 
 from __future__ import annotations
 
-from typing import TYPE_CHECKING
+from typing import TYPE_CHECKING, Any
 
 from torch import nn
 
 from probly.layers.torch import SNGPLayer, SpectralNormWithMultiplier
+from probly.representation.distribution._common import create_categorical_distribution_from_logits
+from probly.representation.distribution.torch_categorical import TorchCategoricalDistributionSample
+from probly.representation.sample.torch import TorchSample
 
 from ._common import (
     EPS,
@@ -17,6 +20,7 @@ from ._common import (
     NORM_MULTIPLIER,
     NUM_INDUCING,
     RIDGE_PENALTY,
+    compute_categorical_sample_from_logits,
     sngp_traverser,
 )
 
@@ -69,3 +73,18 @@ def remove_layer(obj: nn.Softmax, state: State) -> tuple[nn.Module, State]:
     if state[LAST_LAYER]:
         return nn.Identity(), state
     return obj, state
+
+
+@compute_categorical_sample_from_logits.register(TorchSample)
+def torch_compute_categorical_sample_from_logits(
+    sample: TorchSample[Any],
+) -> TorchCategoricalDistributionSample:
+    """Convert a TorchSample of SNGP logits to a categorical distribution sample."""
+    tensor = sample.tensor
+    sample_dim = sample.sample_dim
+    if tensor.ndim >= 3 and sample_dim == 0:
+        tensor = tensor.transpose(0, 1)
+        sample_dim = 1
+
+    categorical_dist = create_categorical_distribution_from_logits(tensor)
+    return TorchCategoricalDistributionSample(tensor=categorical_dist, sample_dim=sample_dim)  # ty: ignore[invalid-argument-type]
