+++ src/probly/representer/__init__.py
@@ -8,7 +8,6 @@ from .credal_set import (
     compute_representative_sample,
 )
 from .sampler import IterableSampler, Sampler
-from .sngp import SNGPRepresenter
 
 __all__ = [
     "ConvexCredalSetRepresenter",
@@ -17,7 +16,6 @@ __all__ = [
     "ProbabilityIntervalsRepresenter",
     "RepresentativeConvexCredalSetRepresenter",
     "Representer",
-    "SNGPRepresenter",
     "Sampler",
     "compute_representative_sample",
     "representer",
